[ISSUE #10289]🐛Move admin SDK error conversion into the adapter layer
